Demande de confirmation de la suppression du compte

+++ src/app/navbar/navbar.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router, RouterLink } from '@angular/router';
 import { AuthService } from '../security/auth.service';
 import { ManageAccountService } from '../security/manage-account.service';
+import { User } from '../models/user';
 
 @Component({
   selector: 'app-navbar',
@@ -11,6 +12,7 @@ import { ManageAccountService } from '../security/manage-account.service';
 export class NavbarComponent implements OnInit {
 
   userAuthenticated : boolean;
+  user : User;
 
   constructor(
     private auth: AuthService, 
@@ -24,13 +26,18 @@ export class NavbarComponent implements OnInit {
         console.log(error);
       }
     });
+
+    this.auth.getUser().subscribe({
+      next: (user) => this.user = user
+    })
   }
   
   deleteAccount(): void {
-    this.accountManager.deleteUser();
-    this.auth.logout();
-    this.router.navigateByUrl("/login");
-
+    if(confirm("Do you really want to delete your account?")){
+      this.accountManager.deleteUser();
+      this.auth.logout();
+      this.router.navigateByUrl("/login");
+    }
   }
 
   ngOnInit(): void {
